refactor: shadcn has been moved to shared/lib
