+++ src/pages/main/ui/Main.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { Link } from '@tanstack/react-router';
 
 import { api } from '../../../shared/api';
-import { Button, Spinner } from '../../../shared/shadcn';
+import { Button, Spinner } from '../../../shared/lib/shadcn';
 
 import type { FC } from 'react';
