+++ vite.config.ts
@@ -1,8 +1,9 @@
-import tailwindcss from "@tailwindcss/vite"
-import { tanstackRouter } from '@tanstack/router-plugin/vite'
-import react from '@vitejs/plugin-react-swc'
-import path from "path"
-import { defineConfig } from 'vite'
+import path from 'path';
+
+import tailwindcss from '@tailwindcss/vite';
+import { tanstackRouter } from '@tanstack/router-plugin/vite';
+import react from '@vitejs/plugin-react-swc';
+import { defineConfig } from 'vite';
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -14,11 +15,11 @@ export default defineConfig({
       routesDirectory: './src/app/router/routes',
     }),
     react(),
-    tailwindcss()
+    tailwindcss(),
   ],
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "./src"),
+      '@': path.resolve(__dirname, './src'),
     },
   },
-})
+});
